pyproject.toml
[project]
name = "aegis-torture"
version = "0.1.0"
description = "AI-powered daily coding exam system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
